refactor(types): use *.ts format for non-autogenned definitions
Using *.ts over *.d.ts allows us to typecheck most of our type libraries
while keeping skipLibCheck set to true
