--- src/types/index.d.ts
@@ -1,5 +1,0 @@
-export type * from './controllers.d.ts';
-export type * from './elements.d.ts';
-export type * from './middlewares.d.ts';
-export type * from './models.d.ts';
-export type * from './validators.d.ts';
