+++ src/types/index.ts
@@ -1,0 +1,5 @@
+export type * from './controllers.ts';
+export type * from './elements.d.ts';
+export type * from './middlewares.ts';
+export type * from './models.d.ts';
+export type * from './validators.ts';
