+++ src/types/validators.ts
@@ -1,7 +1,7 @@
 import type { SchemaObject } from 'ajv';
 import type { IncomingHttpHeaders } from 'node:http';
 
-import type { Record as PiesRecord, RecordLinkage } from './elements.d.ts';
+import type { Record as PiesRecord, RecordLinkage } from './elements.js';
 
 /** Type for validating the integrity of an object's properties. */
 export type IntegrityValidator<T> = {
