+++ tsconfig.json
@@ -20,7 +20,7 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "paths": {
-      "#types": ["./src/types/index.d.ts"],
+      "#types": ["./src/types/index.ts"],
       "#src/*": ["./src/*"]
     },
     "resolveJsonModule": true,
